Separate read-only and editable form hierarchies (#1038)
