+++ collect_app/src/main/java/org/odk/collect/android/activities/ViewFormHierarchyActivity.java
@@ -1,0 +1,125 @@
+/*
+ * Copyright (C) 2017 Shobhit
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package org.odk.collect.android.activities;
+
+
+import android.os.Bundle;
+import android.support.v4.content.ContextCompat;
+import android.view.View;
+import android.widget.AdapterView;
+import android.widget.Button;
+
+import org.javarosa.core.model.FormIndex;
+import org.odk.collect.android.R;
+import org.odk.collect.android.adapters.HierarchyListAdapter;
+import org.odk.collect.android.application.Collect;
+import org.odk.collect.android.exception.JavaRosaException;
+import org.odk.collect.android.logic.HierarchyElement;
+
+import java.util.ArrayList;
+
+import timber.log.Timber;
+
+
+public class ViewFormHierarchyActivity extends FormHierarchyActivity {
+
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        Collect.getInstance().getFormController().stepToOuterScreenEvent();
+
+        Button exitButton = (Button) findViewById(R.id.exitButton);
+        exitButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Collect.getInstance().getActivityLogger().logInstanceAction(this, "exit",
+                        "click");
+                setResult(RESULT_OK);
+                finish();
+            }
+        });
+
+        exitButton.setVisibility(View.VISIBLE);
+
+        jumpBeginningButton.setVisibility(View.GONE);
+        jumpEndButton.setVisibility(View.GONE);
+    }
+
+
+    @Override
+    public void onItemClick(AdapterView<?> parent, View v, int position, long id) {
+        HierarchyElement h = (HierarchyElement) listView.getItemAtPosition(position);
+        FormIndex index = h.getFormIndex();
+        if (index == null) {
+            goUpLevel();
+            return;
+        }
+
+        switch (h.getType()) {
+            case EXPANDED:
+                Collect.getInstance().getActivityLogger().logInstanceAction(this, "onListItemClick",
+                        "COLLAPSED", h.getFormIndex());
+                h.setType(COLLAPSED);
+                ArrayList<HierarchyElement> children = h.getChildren();
+                for (int i = 0; i < children.size(); i++) {
+                    formList.remove(position + 1);
+                }
+                h.setIcon(ContextCompat.getDrawable(getApplicationContext(), R.drawable.expander_ic_minimized));
+                break;
+            case COLLAPSED:
+                Collect.getInstance().getActivityLogger().logInstanceAction(this, "onListItemClick",
+                        "EXPANDED", h.getFormIndex());
+                h.setType(EXPANDED);
+                ArrayList<HierarchyElement> children1 = h.getChildren();
+                for (int i = 0; i < children1.size(); i++) {
+                    Timber.i("adding child: %s", children1.get(i).getFormIndex());
+                    formList.add(position + 1 + i, children1.get(i));
+
+                }
+                h.setIcon(ContextCompat.getDrawable(getApplicationContext(), R.drawable.expander_ic_maximized));
+                break;
+            case QUESTION:
+                Collect.getInstance().getActivityLogger().logInstanceAction(this, "onListItemClick",
+                        "QUESTION-JUMP", index);
+                Collect.getInstance().getFormController().jumpToIndex(index);
+                if (Collect.getInstance().getFormController().indexIsInFieldList()) {
+                    try {
+                        Collect.getInstance().getFormController().stepToPreviousScreenEvent();
+                    } catch (JavaRosaException e) {
+                        Timber.e(e);
+                        createErrorDialog(e.getCause().getMessage());
+                        return;
+                    }
+                }
+                setResult(RESULT_OK);
+                return;
+            case CHILD:
+                Collect.getInstance().getActivityLogger().logInstanceAction(this, "onListItemClick",
+                        "REPEAT-JUMP", h.getFormIndex());
+                Collect.getInstance().getFormController().jumpToIndex(h.getFormIndex());
+                setResult(RESULT_OK);
+                refreshView();
+                return;
+        }
+
+        // Should only get here if we've expanded or collapsed a group
+        HierarchyListAdapter itla = new HierarchyListAdapter(this);
+        itla.setListItems(formList);
+        listView.setAdapter(itla);
+        listView.setSelection(position);
+    }
+}
